Refactor rename gui to view to follow the Glimmer app conventions for where Views live
