+++ lib/adamantite.rb
@@ -12,13 +12,13 @@ require 'model/request/add_password_request'
 require 'model/request/update_master_password_request'
 require 'model/request/set_master_license_key_request'
 require 'model/request/set_master_password_request'
-require 'gui/screen/copy_screen'
-require 'gui/screen/login_screen'
-require 'gui/screen/set_master_password_screen'
-require 'gui/screen/show_screen'
-require 'gui/screen/update_master_password_screen'
-require 'gui/form/license_form'
-require 'gui/form/password_object_form_window'
+require 'view/screen/copy_screen'
+require 'view/screen/login_screen'
+require 'view/screen/set_master_password_screen'
+require 'view/screen/show_screen'
+require 'view/screen/update_master_password_screen'
+require 'view/form/license_form'
+require 'view/form/password_object_form_window'
 
 module Adamantite
   class AdamantiteApp
