+++ lib/view/form/license_form.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Adamantite
-  module GUI
+  module View
     module Screen
       class LicenseForm
         include Glimmer::LibUI::CustomWindow
